% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower.R
\name{draw_flowers}
\alias{draw_flowers}
\title{Draw flowers}
\usage{
draw_flowers(
  my_df,
  my_grouping_var,
  my_metric_var,
  my_color_var = NULL,
  my_x_margin = 0,
  my_y_jitter = 0,
  my_curvature = 0.5,
  my_angle = 90
)
}
\arguments{
\item{my_df}{data frame}

\item{my_grouping_var}{field within data frame}

\item{my_metric_var}{numeric field within data frame}

\item{my_color_var}{field within data frame}

\item{my_x_margin}{numeric variable}

\item{my_y_jitter}{numeric variable}

\item{my_curvature}{numeric variable}

\item{my_angle}{numeric variable between 0 and 180}
}
\description{
Draw flowers
}
